requirement.txt
APScheduler==3.6.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-APScheduler==1.11.0
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
PyMySQL==0.9.3
pypinyin==0.37.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
spyne==2.12.16
SQLAlchemy==1.3.13
suds-jurko==0.6
tzlocal==2.0.0
urllib3==1.25.8
virtualenv==16.7.9
Werkzeug==1.0.0
xmltodict==0.12.0
